+++ src/library/DIPS.Mobile.UI/Components/Labels/Android/MauiTextView.cs
@@ -1,6 +1,5 @@
 using Android.Content;
 using Android.Graphics;
-using Android.Text;
 using Android.Text.Style;
 using DIPS.Mobile.UI.Resources.LocalizedStrings.LocalizedStrings;
 using Kotlin.Text;
@@ -12,7 +11,6 @@ namespace DIPS.Mobile.UI.Components.Labels.Android;
 public class MauiTextView : Microsoft.Maui.Platform.MauiTextView
 {
     private readonly Label m_label;
-    private readonly SpannableString m_newEllipsis;
 
     public MauiTextView(Context context, Label label) : base(context)
     {
@@ -33,22 +31,8 @@ public class MauiTextView : Microsoft.Maui.Platform.MauiTextView
         {
             text = m_label.Text;
         }
-        
-        var availableViewWidth = MeasuredWidth - (float)CompoundPaddingLeft - CompoundPaddingRight;
-        var availableTextWidth = availableViewWidth * MaxLines;
 
-        var ellipsizedText = TextUtils.Ellipsize(text, Paint, availableTextWidth, Ellipsize);
-
-        if (ellipsizedText != text)
-        {
-            m_label.IsEllipsized = true;
-        }
-        else
-        {
-            m_label.IsEllipsized = false;
-        }
+        m_label.IsTruncated = Layout?.Text != text;
     }
-
-   
 
 }