+++ src/library/DIPS.Mobile.UI/Components/TextFields/Editor/EditorHandler.cs
@@ -9,9 +9,11 @@ public partial class EditorHandler : Microsoft.Maui.Handlers.EditorHandler
     public static IPropertyMapper<Editor, EditorHandler> PropertyMapper = new PropertyMapper<Editor, EditorHandler>(Mapper)
     {
         [nameof(Editor.HasBorder)] = MapHasBorder,
-        [nameof(Editor.ShouldSelectAllTextOnFocused)] = MapShouldSelectTextOnTapped
+        [nameof(Editor.ShouldSelectAllTextOnFocused)] = MapShouldSelectTextOnTapped,
+        [nameof(Editor.ShouldUseDefaultPadding)] = MapShouldUseDefaultPadding
     };
 
+    private static partial void MapShouldUseDefaultPadding(EditorHandler handler, Editor editor);
     private static partial void MapShouldSelectTextOnTapped(EditorHandler handler, Editor entry);
     private static partial void MapHasBorder(EditorHandler handler, Editor entry);
 }