+++ src/library/DIPS.Mobile.UI/Components/TextFields/Editor/iOS/EditorHandler.cs
@@ -1,4 +1,5 @@
 using Microsoft.Maui.Platform;
+using UIKit;
 
 namespace DIPS.Mobile.UI.Components.TextFields.Editor;
 
@@ -11,6 +12,16 @@ public partial class EditorHandler
         platformView.VerticalTextAlignment = TextAlignment.Start;
         platformView.Started += OnFocus;
     }
+
+    private static partial void MapShouldUseDefaultPadding(EditorHandler handler, Editor editor)
+    {
+        if(editor.ShouldUseDefaultPadding)
+            return;
+
+        handler.PlatformView.TextContainer.LineFragmentPadding = 0;
+        handler.PlatformView.TextContainerInset = UIEdgeInsets.Zero;
+    }
+
 
     private async void OnFocus(object? sender, EventArgs e)
     {