+++ src/library/DIPS.Mobile.UI/Components/TextFields/Entry/dotnet/EntryHandler.cs
@@ -11,4 +11,9 @@ public partial class EntryHandler
     {
         
     }
+    
+    private static partial void MapShouldUseDefaultPadding(EntryHandler handler, Entry entry)
+    {
+        
+    }
 }