+++ src/library/DIPS.Mobile.UI/Resources/Styles/InputField/InputFieldStyle.cs
@@ -1,0 +1,9 @@
+namespace DIPS.Mobile.UI.Resources.Styles.InputField;
+
+public enum InputFieldStyle
+{
+    None=0,
+    Default=1,
+    Focused=2,
+    Disabled=3
+}